requirements.txt
nltk==3.7
pandas==1.5.0
praw==7.6.1
python-dotenv==0.21.0
psaw==0.1.0
psycopg2==2.9.5
scipy==1.9.2
transformers==4.24.0
